+++ tapo/src/responses/device_info_result.rs
@@ -1,7 +1,7 @@
+mod basic;
 mod color_light;
 mod default_plug_state;
 mod default_state;
-mod generic;
 mod hub;
 mod light;
 mod plug;
@@ -11,10 +11,10 @@ mod power_strip;
 mod rgb_light_strip;
 mod rgbic_light_strip;
 
+pub use basic::*;
 pub use color_light::*;
 pub use default_plug_state::*;
 pub use default_state::*;
-pub use generic::*;
 pub use hub::*;
 pub use light::*;
 pub use plug::*;