+++ tapo/src/responses/device_info_result/power_strip.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 use crate::error::Error;
 use crate::responses::{DecodableResultExt, TapoResponseExt, decode_value};
 
-/// Device info of Tapo P300, P304M, P306 and P316M. Superset of [`crate::responses::DeviceInfoGenericResult`].
+/// Device info of Tapo P300, P304M, P306 and P316M.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[cfg_attr(feature = "python", pyo3::prelude::pyclass(from_py_object, get_all))]
 #[allow(missing_docs)]
 pub struct DeviceInfoPowerStripResult {
     //
-    // Inherited from DeviceInfoGenericResult
+    // Common properties
     //
     pub avatar: String,
     pub device_id: String,